% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnROC.R
\name{bnROC}
\alias{bnROC}
\title{The bnROC Function}
\usage{
bnROC(score, status)
}
\arguments{
\item{score:}{a vector of test scores, continuous}

\item{status:}{a vector of disease statuses, binary (0/1)}
}
\value{
A list containing important variables like AUC, TPR, and FPR.
}
\description{
This function computes the ROC curve and the AUC value
for a range of cutoff thresholds between \link{-3,3}.
}
