% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnROC.R
\name{MLestimates}
\alias{MLestimates}
\title{ML Estimates Function
Computes the mean and standard deviation of the input vector}
\usage{
MLestimates(x)
}
\arguments{
\item{x:}{a vector of numbers}
}
\value{
a list containing the mean and standard deviation.
}
\description{
ML Estimates Function
Computes the mean and standard deviation of the input vector
}
